split test_models too
